Asynchronous Recording This is a fix for bug250 - Recorded files plays too fast.
RCA
We were doing the record operation on the grabbing thread, slowing it down.


Implemented
- On frame grab, switch back to main thread (=UI thread) as fast as possible.
- Creation of a producer-consumer class for recording. Spawns a new thread and push frames in a queue. The queue is then processed by this new recording thread.
- Removed framerate estimation concept as the WinForms timer is not reliable, and we are now asynchronous so we honor the actual framerate.
- Removed dropping and flickering protection as they don't fit in the new paradigm. Will have to implement queue overflow protection.
- Update to status bar is only requested when the buffer fill value change, but now it's updated in real time during refill.
- Gave names to some threads (main, grab, record, thumb loaders)
- Update to Aforge SVN r1538.

+++ ScreenManager/PlayerScreen/VideoFilters/VideoFilterContrast.cs
@@ -115,7 +115,7 @@ namespace Kinovea.ScreenManager
 		}
 		private Bitmap ProcessSingleImage(Bitmap _src)
 		{
-			double fValue = 1.6;
+			float fValue = 1.6F;
 			Bitmap img = (_src.PixelFormat == PixelFormat.Format24bppRgb) ? _src : CloneTo24bpp(_src);
 			
 			ContrastCorrection filter = new ContrastCorrection(fValue);
